Remove composeApp (ri ☠)
