--- composeApp/src/androidMain/kotlin/io/rwc/streamwise/MainActivity.kt
@@ -1,25 +1,0 @@
-package io.rwc.streamwise
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-
-class MainActivity : ComponentActivity() {
-  override fun onCreate(savedInstanceState: Bundle?) {
-    enableEdgeToEdge()
-    super.onCreate(savedInstanceState)
-
-    setContent {
-      App()
-    }
-  }
-}
-
-@Preview
-@Composable
-fun AppAndroidPreview() {
-  App()
-}
