--- composeApp/src/commonTest/kotlin/io/rwc/streamwise/ComposeAppCommonTest.kt
@@ -1,12 +1,0 @@
-package io.rwc.streamwise
-
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class ComposeAppCommonTest {
-
-  @Test
-  fun example() {
-    assertEquals(3, 1 + 2)
-  }
-}
