--- composeApp/src/desktopMain/kotlin/io/rwc/streamwise/Main.kt
@@ -1,13 +1,0 @@
-package io.rwc.streamwise
-
-import androidx.compose.ui.window.Window
-import androidx.compose.ui.window.application
-
-fun main() = application {
-  Window(
-    onCloseRequest = ::exitApplication,
-    title = "Streamwise",
-  ) {
-    App()
-  }
-}
