--- composeApp/src/wasmJsMain/kotlin/io/rwc/streamwise/main.kt
@@ -1,12 +1,0 @@
-package io.rwc.streamwise
-
-import androidx.compose.ui.ExperimentalComposeUiApi
-import androidx.compose.ui.window.ComposeViewport
-import kotlinx.browser.document
-
-@OptIn(ExperimentalComposeUiApi::class)
-fun main() {
-  ComposeViewport(document.body!!) {
-    App()
-  }
-}
