--- composeApp/src/wasmJsMain/resources/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>streamwise</title>
-  <link type="text/css" rel="stylesheet" href="styles.css">
-  <script type="application/javascript" src="composeApp.js"></script>
-</head>
-<body>
-</body>
-</html>
