+++ settings.gradle.kts
@@ -32,6 +32,5 @@ plugins {
   id("org.gradle.toolchains.foojay-resolver-convention") version "0.10.0"
 }
 
-include(":composeApp")
 include(":server")
 include(":shared")
